+++ tests/anomaly_detection/test_detectors.py
@@ -22,9 +22,11 @@ DETECTORS_NOT_MULTIVARIATE = [
     anomaly_detection.baselines.AlwaysNormal(),
     anomaly_detection.baselines.AlwaysAnomalous(),
     anomaly_detection.baselines.RandomDetector(seed=42),
+    anomaly_detection.ClusterBasedLocalOutlierFactor(15),
     anomaly_detection.HistogramBasedOutlierScore(1),
     anomaly_detection.IsolationForest(15),
     anomaly_detection.KernelPrincipalComponentAnalysis(15),
+    anomaly_detection.KMeansAnomalyDetector(15),
     anomaly_detection.KNearestNeighbors(15),
     anomaly_detection.LocalOutlierFactor(15),
     anomaly_detection.MatrixProfileDetector(15, novelty=False),
@@ -112,9 +114,11 @@ class TestAnomalyDetectors:
 
 
 @pytest.mark.parametrize('detector_class,additional_args', [
+    (anomaly_detection.ClusterBasedLocalOutlierFactor, {'n_clusters': 20}),
     (anomaly_detection.HistogramBasedOutlierScore, {}),
     (anomaly_detection.IsolationForest, {}),
     (anomaly_detection.KernelPrincipalComponentAnalysis, {}),
+    (anomaly_detection.KMeansAnomalyDetector, {}),
     (anomaly_detection.KNearestNeighbors, {}),
     (anomaly_detection.LocalOutlierFactor, {}),
     (anomaly_detection.MatrixProfileDetector, {}),