+++ docs/additional_information/changelog.rst
@@ -9,6 +9,8 @@ Latest
 
 Added
 ^^^^^
+- Implemented ``ClusterBasedLocalOutlierFactor`` (CBLOF) anomaly detector.
+- Implemented ``KMeansAnomalyDetector`` anomaly detector.
 
 Changed
 ^^^^^^^